docker/backend/image/pegbbackend/requirements.txt
Django
asgiref
django-rest-framework
djangorestframework
pip
setuptools
sqlparse
tzdata
wheel
mysqlclient
djangorestframework-simplejwt